requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
itsdangerous
python-dotenv
aiofiles
email-validator
